SocialLLM/requirements.txt
numpy>=1.20.0
pyyaml>=5.4.0
openai>=1.0.0
tenacity>=8.0.0
mesa>=2.0.0
tqdm>=4.60.0